+++ src/index.ts
@@ -28,6 +28,13 @@ export {
 } from "./Domain/Interfaces";
 export { JsonConverter, JsonSerializer, JsonSerializerOptions } from "./System/Text/Json";
 export * from "./System/Net/Http";
+export * from "./System/Net/Http";
+export { AnsiConsole } from "./System/Console/AnsiConsole";
+export { Color } from "./System/Console/Color";
+export { Rule } from "./System/Console/Rule";
+export { Panel } from "./System/Console/Panel";
+export { Table } from "./System/Console/Table";
+
 
 // Collections (Explicit exports due to missing index)
 export * from "./System/Collections/Generic/List";